+++ tutorial/Task7Group_InSight/src/utility.h
@@ -29,8 +29,8 @@ bool read_cnn_weights_file_char(char *filename , std::vector<std::vector<std::ve
 void ReadMNIST_char(int NumberOfImages,int NumberOfRows,int NumberOfCols,int ZERO_PADDING,std::vector<std::vector<std::vector<unsigned char>>> &arr);
 
 void convlutionLayer(std::vector<std::vector<unsigned char>> &ImageReader,std::vector<std::vector<short>> &CNNWeights,short cnnbias,int FILTER_ROWS,int FILTER_COLS,
-  int NUMBER_OF_ROWS,int NUMBER_OF_COLS,std::vector<std::vector<long>> &ConvOutput, int CONV_LAYER_OUTPUT_ROWS, int CONV_LAYER_OUTPUT_COLS);
+  int NUMBER_OF_ROWS,int NUMBER_OF_COLS,std::vector<std::vector<int>> &ConvOutput, int CONV_LAYER_OUTPUT_ROWS, int CONV_LAYER_OUTPUT_COLS);
 
-void maxpoolLayer(std::vector<std::vector<std::vector<long>>> &ConvOutputFilters,std::vector<std::vector<std::vector<long>>> &MaxPoolOutput,int NUMBER_OF_FILTERS,int NUMBER_OF_ROWS,int NUMBER_OF_COLS,int STRIDE);
-int fullyConnectedLayer(long *MaxPoolOutput_1D,short *weight,int numberOfFCPixels);
+void maxpoolLayer(std::vector<std::vector<std::vector<int>>> &ConvOutputFilters,std::vector<std::vector<std::vector<int>>> &MaxPoolOutput,int NUMBER_OF_FILTERS,int NUMBER_OF_ROWS,int NUMBER_OF_COLS,int STRIDE);
+int fullyConnectedLayer(int *MaxPoolOutput_1D,short *weight,int numberOfFCPixels);
 #endif
